+++ Backend/Interfaces/Word/IWordService.cs
@@ -2,5 +2,6 @@
 {
     public interface IWordService
     {
+        Task<List<Word>> GetWordsByWordTypeId(int wordTypeIsd);
     }
 }