+++ Changelog.md
@@ -67,3 +67,9 @@ All notable changes to this project will be documented in this file.
 
 ## Fixed
 - Fixed DatabaSeeder to all camelcases data and auto-increment ID's
+
+
+  ### [0.0.3] - 2026-08-05
+
+  ### Added 
+  - Added word functionality and endpoint